+++ build.sh
@@ -15,6 +15,9 @@ DC_COMPANY_ID=au.com.derekclarkson
 DC_AUTHOR="Derek Clarkson"
 DC_COMPANY=$DC_AUTHOR
 
+DC_SIMULATOR_SDK=iphonesimulator4.3
+DC_DEVICE_SDK=iphoneos4.3
+
 DC_SCRIPTS_DIR=scripts
 
 # Include common scripts.